pyproject.toml
[project]
name = "spider"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.10"
dependencies = [
    "fastapi>=0.122.0",
    "pytest-playwright>=0.7.2",
    "ruff>=0.14.6",
    "supabase>=2.0.0",
    "python-dotenv>=1.0.0",
    "openpyxl>=3.1.2",
]

#[tool.setuptools]
## 自动发现 src 目录下的所有包
#packages = {find = {where = ["src"]}}
#
#[tool.setuptools.package-dir]
#"" = "src"
